Replacing the "connect" buttons to a "Connecting..." screen, and lots of change to make an actual Flone fly

+++ app/src/main/java/com/fewlaps/flone/io/input/UserInstructionsInput.java
@@ -10,7 +10,7 @@ package com.fewlaps.flone.io.input;
  * @date 14/05/2015
  */
 public interface UserInstructionsInput {
-    public int getThrottle();
+    public double getThrottle();
 
     public double getHeading();
